Refactored ModelScreenshotter
