+++ example/web-avatar-client/package.json
@@ -17,7 +17,6 @@
   "dependencies": {
     "@mml-io/3d-web-client-core": "^0.11.0",
     "@mml-io/3d-web-standalone-avatar-editor": "^0.11.0",
-    "@mml-io/3d-web-model-screenshot": "^0.11.0",
     "three": "0.153.0"
   },
   "devDependencies": {
