--- packages/3d-web-model-screenshot/build.ts
@@ -1,3 +1,0 @@
-import { handleLibraryBuild } from "../../utils/build-library";
-
-handleLibraryBuild();
