--- packages/3d-web-model-screenshot/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "@mml-io/3d-web-model-screenshot",
-  "version": "0.11.0",
-  "publishConfig": {
-    "access": "public"
-  },
-  "main": "./build/index.js",
-  "types": "./build/index.d.ts",
-  "type": "module",
-  "files": [
-    "/build"
-  ],
-  "scripts": {
-    "build": "tsx ./build.ts --build",
-    "iterate": "tsx ./build.ts --watch",
-    "type-check": "tsc --noEmit",
-    "lint": "eslint \"./{src,test}/**/*.{js,jsx,ts,tsx}\" --max-warnings 0",
-    "lint-fix": "eslint \"./{src,test}/**/*.{js,jsx,ts,tsx}\" --fix"
-  },
-  "dependencies": {
-    "@mml-io/3d-web-avatar": "^0.11.0"
-  },
-  "peerDependencies": {
-    "three": ">= 0.153.0"
-  },
-  "devDependencies": {
-    "@types/three": "0.153.0"
-  }
-}
