--- packages/3d-web-model-screenshot/tsconfig.json
@@ -1,22 +1,0 @@
-{
-  "compilerOptions": {
-    "noImplicitAny": true,
-    "module": "esnext",
-    "target": "ES6",
-    "jsx": "react",
-    "lib": ["es2020", "dom"],
-    "esModuleInterop": true,
-    "sourceMap": true,
-    "allowJs": true,
-    "types": ["node", "jest"],
-    "strictNullChecks": true,
-    "allowSyntheticDefaultImports": true,
-    "moduleResolution": "node",
-    "resolveJsonModule": true,
-    "downlevelIteration": true,
-    "outDir": "./build",
-    "incremental": true
-  },
-  "include": ["src/**/*", "test/**/*", "./build.ts"],
-  "exclude": ["**/build/*", "types-src"]
-}
