+++ packages/3d-web-standalone-avatar-editor/src/scene/Lights.ts
@@ -1,14 +1,13 @@
-import { AmbientLight, DirectionalLight, OrthographicCamera, Vector3 } from "three";
+import { AmbientLight, DirectionalLight, Group, OrthographicCamera, Vector3 } from "three";
 
-export class Lights {
+export class Lights extends Group {
   public ambientLight: AmbientLight;
   public frontLight: DirectionalLight;
   public backLight: DirectionalLight;
-  public frontLightC: DirectionalLight;
 
   private shadowResolution: number = 8192;
   private shadowCamFrustum: number = 10;
-  private lookAt: Vector3;
+  private lookAtTarget: Vector3;
 
   private shadowCamera = new OrthographicCamera(
     -this.shadowCamFrustum,
@@ -20,9 +19,11 @@ export class Lights {
   );
 
   constructor(private cameraOffset: Vector3) {
-    this.lookAt = new Vector3().copy(new Vector3()).add(this.cameraOffset);
+    super();
+    this.lookAtTarget = new Vector3().copy(new Vector3()).add(this.cameraOffset);
 
     this.ambientLight = new AmbientLight(0xffffff, 0.4);
+    this.add(this.ambientLight);
 
     this.frontLight = new DirectionalLight(0xffeedd, 1.0);
     this.frontLight.position.set(2, 4, 2);
@@ -31,7 +32,8 @@ export class Lights {
     this.frontLight.shadow.camera = this.shadowCamera;
     this.frontLight.shadow.mapSize.set(this.shadowResolution, this.shadowResolution);
     this.frontLight.castShadow = true;
-    this.frontLight.target.position.copy(this.lookAt);
+    this.frontLight.target.position.copy(this.lookAtTarget);
+    this.add(this.frontLight);
 
     this.backLight = new DirectionalLight(0xddffff, 2.0);
     this.backLight.position.set(2, 4, -3);
@@ -40,7 +42,8 @@ export class Lights {
     this.backLight.shadow.camera = this.shadowCamera;
     this.backLight.shadow.mapSize.set(this.shadowResolution, this.shadowResolution);
     this.backLight.castShadow = true;
-    this.backLight.target.position.copy(this.lookAt);
+    this.backLight.target.position.copy(this.lookAtTarget);
+    this.add(this.backLight);
 
     this.frontLight.target.updateMatrixWorld();
     this.backLight.target.updateMatrixWorld();
